Move `gamepadAction` handler to its own file
